requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
python-wordpress-xmlrpc>=2.3
schedule>=1.1.0

